requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-lru==2.0.4
attrs==23.2.0
ddapi==0.4.7
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
nextcord==2.6.0
numpy==2.0.0
orjson==3.10.6
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2rc1
typing_extensions==4.12.2
yarl==1.9.4
